+++ lib/ext/solid/auth/interfaces/solid_auth_operations.dart
@@ -3,15 +3,12 @@ import 'package:solid_task/ext/solid_flutter/auth/integration/solid_authenticati
 
 /// Interface defining operations for SOLID authentication
 abstract class SolidAuthOperations<C> {
-  /// Computes the issuer URI from a WebId URI, issuer URI, or Pod URL
-  Future<String> getIssuer(String input);
-
   /// Authenticates a user with a SOLID identity provider
   ///
-  /// [issuerUri] is the URI of the SOLID identity provider
+  /// [webIdOrIssuerUri] is the WebID or the URI of the SOLID identity provider
   /// [context] is the Flutter build context needed for web authentication flows
   /// Returns an [AuthResult] containing authentication data or an error
-  Future<AuthResult> authenticate(String issuerUri, C context);
+  Future<AuthResult> authenticate(String webIdOrIssuerUri, C context);
 
   /// Logs the current user out
   ///