+++ test/mocks/integration_test_mocks.mocks.dart
@@ -6,9 +6,8 @@
 import 'dart:async' as _i7;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i8;
 import 'package:solid_task/ext/solid/auth/interfaces/auth_state_change_provider.dart'
-    as _i9;
+    as _i8;
 import 'package:solid_task/ext/solid/auth/interfaces/solid_auth_operations.dart'
     as _i6;
 import 'package:solid_task/ext/solid/auth/interfaces/solid_auth_state.dart'
@@ -17,7 +16,7 @@ import 'package:solid_task/ext/solid/auth/models/auth_result.dart' as _i2;
 import 'package:solid_task/ext/solid/sync/sync_service.dart' as _i4;
 import 'package:solid_task/ext/solid_flutter/auth/integration/solid_authentication_backend.dart'
     as _i3;
-import 'package:solid_task/services/logger_service.dart' as _i10;
+import 'package:solid_task/services/logger_service.dart' as _i9;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -88,30 +87,15 @@ class MockSolidAuthOperations<C> extends _i1.Mock
   }
 
   @override
-  _i7.Future<String> getIssuer(String? input) => (super.noSuchMethod(
-        Invocation.method(
-          #getIssuer,
-          [input],
-        ),
-        returnValue: _i7.Future<String>.value(_i8.dummyValue<String>(
-          this,
-          Invocation.method(
-            #getIssuer,
-            [input],
-          ),
-        )),
-      ) as _i7.Future<String>);
-
-  @override
   _i7.Future<_i2.AuthResult> authenticate(
-    String? issuerUri,
+    String? webIdOrIssuerUri,
     C? context,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
           #authenticate,
           [
-            issuerUri,
+            webIdOrIssuerUri,
             context,
           ],
         ),
@@ -120,7 +104,7 @@ class MockSolidAuthOperations<C> extends _i1.Mock
           Invocation.method(
             #authenticate,
             [
-              issuerUri,
+              webIdOrIssuerUri,
               context,
             ],
           ),
@@ -176,7 +160,7 @@ class MockSolidAuthOperations<C> extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockAuthStateChangeProvider extends _i1.Mock
-    implements _i9.AuthStateChangeProvider {
+    implements _i8.AuthStateChangeProvider {
   MockAuthStateChangeProvider() {
     _i1.throwOnMissingStub(this);
   }
@@ -278,7 +262,7 @@ class MockSyncService extends _i1.Mock implements _i4.SyncService {
 /// A class which mocks [ContextLogger].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockContextLogger extends _i1.Mock implements _i10.ContextLogger {
+class MockContextLogger extends _i1.Mock implements _i9.ContextLogger {
   MockContextLogger() {
     _i1.throwOnMissingStub(this);
   }